Create L8C1
Briefing
New Line of Communication
When you were on the last level we discovered the aliens had been communicating with a human insider back on Earth, and by email of all things! Well, it seems that may have been a bit of a decoy as they're also exchanging secret encrypted messages from a server on one of the probe ships.

We need you to write a script which can connect over TCP to the following server - localhost:10000 and send 'GET' to retrieve the encrypted messages. You then need to decrypt them and send them back split by \n. We think you'll receive 3 sentences, and that they're encrypted with a caesar cipher with a different random offset per sentence.

Finally, print the response after sending the decrypted messages to get the flag.

Tip: Remember, each encrypted message has a different random offset.